% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monashthesisdown.R, R/thesisdown.R
\docType{package}
\name{monashthesisdown}
\alias{monashthesisdown}
\alias{monashthesisdown-package}
\title{monashthesisdown: A package for creating undergrad and post-grad theses using RMarkdown}
\description{
This package inspired by thesisdown and  Prof’ Rob J Hyndman’s work, uses bookdown package  knits your thesis in pdf, gitbook and epub format according to the guidelines of The Monash University

This package helps to render your thesis based on Monash Templete.
}
\details{
It is advisable that you load **bookdown**, along with **monashthesisdown**
}
\section{thesis_gitbook}{

Creates an R Markdown thesis template as a webpage
}

\section{thesis_pdf}{

Creates an R Markdown thesis template as a PDF
}

\section{thesis_word}{

Creates an R Markdown thesis template as a Microsoft Word document
}

\section{thesis_epub}{

Creates an R Markdown thesis template as an ebook
}

\author{
\strong{Maintainer}: Mushfiqul Anwar Siraji \email{siraji1993@gmail.com} (\href{https://orcid.org/0000-0003-0127-9982}{ORCID})

}
